+++ src/useHandlers.ts
@@ -13,27 +13,43 @@ export type eventLikeHandlers = Handler<string | EventTarget | null, keyValueEve
  * for calling a list of functions with similar parameters. A great use case is creating
  * a reset form function from a list of reset functions (see example below). In fact, the
  * `reset` function from {@link useForm} is created using this function.
- * @param handlers - the list of specified functions
+ * @param handlers - the list of specified functions.
  * @returns A single function.  When this function is invoked, it will call each specified
  * function in the order it was passed to this hook with the same arguments from the invocation.
+ * If one of the specified handlers returns a promise, then this function will also return a promise
+ * using the Promise.allSettled strategy.
  * @example
  * ```javascript
  * // create a reset form function by merging reset functions from other form hooks.
  * const reset = useHandlers(resetValues, resetErrors, resetTouches);
  * ```
  */
-export function useHandlers<T, K extends T[]>(
-  ...handlers: Handler<T, K, void>[]
-): Handler<T, K, void> {
-  return useCallback((...args: K): void => {
-    handlers.forEach((func): void => {
-      assert.error(
-        typeof func === 'function',
-        LoggingTypes.typeError,
-        `(expected: function, received: ${typeof func})`,
-      );
-      func(...args);
-    });
+export function useHandlers<
+  T,
+  K extends T[],
+  R extends Promise<void> | void = void,
+  Return = R extends void ? void : Promise<void>
+>(...handlers: Handler<T, K, R>[]): Handler<T, K, Return> {
+  // eslint-disable-next-line @typescript-eslint/ban-ts-comment
+  // @ts-expect-error don't know how to fix this
+  // eslint-disable-next-line consistent-return
+  return useCallback((...args: K): Return => {
+    let hasAPromise = false;
+    const voidOrPromises = handlers.map(
+      (func): R => {
+        assert.error(
+          typeof func === 'function',
+          LoggingTypes.typeError,
+          `(expected: function, received: ${typeof func})`,
+        );
+        const res = func(...args);
+        hasAPromise = hasAPromise || res instanceof Promise;
+        return res;
+      },
+    );
+    if (hasAPromise) {
+      return (Promise.allSettled(voidOrPromises) as unknown) as Return;
+    }
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, handlers);
 }