+++ src/utils.ts
@@ -1,3 +1,4 @@
+/* eslint-disable @typescript-eslint/ban-ts-comment */
 import { SyntheticEvent } from 'react';
 
 const projectName = 'react-uniformed';
@@ -13,6 +14,7 @@ interface Logger {
 
 export function resetForm(event?: SyntheticEvent): void {
   if (event?.target instanceof HTMLElement) {
+    // eslint-disable-next-line
     event.target.closest('form')?.reset();
   }
 }