+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- [`useForm`] Fixed an issue where failed submissions would reset the form.
+
 ## [0.0.19] - 2020-07-15
 
 ### Fixed